requirements.txt
streamlit==1.19.0
streamlit-image-comparison==0.0.3
opencv-python-headless
sahi
Pillow==9.4.0
moviepy==1.0.3
pandas==1.5.3
replicate==0.4.0
requests==2.28.2
boto3==1.26.54
ffmpeg-python==0.2.0
streamlit-drawable-canvas==0.9.2
numpy==1.24.1
mesa
streamlit-javascript==0.1.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
streamlit-cropper==0.2.1
pydub==0.25.1
requests-toolbelt==1.0.0
streamlit-option-menu==0.3.4
colorlog==6.7.0
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==0.19.2